Unbreak plain projects
A recent change to make error handling better in React broke plain js
projects that didn't have 'react' specied as a dependency. We need a
separate renderWrapper for plain projects.
