+++ src/DEFAULTS.js
@@ -25,7 +25,7 @@ export const plugins = [];
 export const prerender = true;
 export const publicFolders = [];
 export const tmpdir = path.join(os.tmpdir(), cwdUniqueFolderName);
-export const renderWrapperModule = require.resolve('./renderWrapper');
+export const renderWrapperModule = undefined;
 export const githubApiUrl = 'https://api.github.com';
 export const compareThreshold = undefined;
 export const pages = undefined;
